+++ CHANGELOG.md
@@ -11,6 +11,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 ### Fixed
 - Filter `cdylib` libraries by the `TargetKind` and not by the file extension.
 - *Others* size.
+- Rows count specified by the `-n` flag.
 
 ## [0.5.0] - 2018-05-29
 ### Added